Rename `WaitScene` to `NewAccountWaitScene`
This isn't a general-purpose wait scene - it's very much designed for account creation with the vault animation.
